Implemented font texture caching
This seems to be quite a bit faster and solves the wavy line problems on mac.
I implemented a switch so this can be turned on/off in gGraphView::DrawTextQue()

change use_pixmap_cache to false to turn it off.

I'll remove the graph title and yaxis bits and put them back as this has the
same effect